+++ backend/bin/run-jobs.sh
@@ -3,6 +3,7 @@
 DIR=$(dirname "$(realpath "$0")")
 
 ${DIR}/console update-chars --log --hide-details
+${DIR}/console update-corporations --log --hide-details
 ${DIR}/console update-player-groups --log --hide-details
 ${DIR}/console check-tokens --log --hide-details
 ${DIR}/console update-member-tracking --log