+++ backend/src/Command/UpdateCharacters.php
@@ -6,13 +6,9 @@ namespace Neucore\Command;
 
 use Neucore\Command\Traits\EsiRateLimited;
 use Neucore\Command\Traits\LogOutput;
-use Neucore\Entity\Alliance;
 use Neucore\Entity\Character;
-use Neucore\Entity\Corporation;
 use Neucore\Factory\RepositoryFactory;
-use Neucore\Repository\AllianceRepository;
 use Neucore\Repository\CharacterRepository;
-use Neucore\Repository\CorporationRepository;
 use Neucore\Service\EntityManager;
 use Neucore\Service\EsiData;
 use Psr\Log\LoggerInterface;
@@ -27,24 +23,10 @@ class UpdateCharacters extends Command
     use LogOutput;
     use EsiRateLimited;
 
-    const UPDATE_OK = 'update OK';
-
-    const UPDATE_NOK = 'update NOK';
-
     /**
      * @var CharacterRepository
      */
     private $charRepo;
-
-    /**
-     * @var CorporationRepository
-     */
-    private $corpRepo;
-
-    /**
-     * @var AllianceRepository
-     */
-    private $alliRepo;
 
     /**
      * @var EsiData
@@ -77,8 +59,6 @@ class UpdateCharacters extends Command
         $this->esiRateLimited($repositoryFactory->getSystemVariableRepository());
 
         $this->charRepo = $repositoryFactory->getCharacterRepository();
-        $this->corpRepo = $repositoryFactory->getCorporationRepository();
-        $this->alliRepo = $repositoryFactory->getAllianceRepository();
         $this->esiData = $esiData;
         $this->entityManager = $entityManager;
     }
@@ -86,8 +66,8 @@ class UpdateCharacters extends Command
     protected function configure(): void
     {
         $this->setName('update-chars')
-            ->setDescription('Updates all characters, corporations and alliances from ESI.')
-            ->addArgument('character', InputArgument::OPTIONAL, 'Update one char, no corporations or alliances.')
+            ->setDescription('Updates all characters from ESI.')
+            ->addArgument('character', InputArgument::OPTIONAL, 'Update one char.')
             ->addOption(
                 'sleep',
                 's',
@@ -107,10 +87,6 @@ class UpdateCharacters extends Command
         $this->writeLine('Started "update-chars"', false);
 
         $this->updateChars($charId);
-        if ($charId === 0) {
-            $this->updateCorps();
-            $this->updateAlliances();
-        }
 
         $this->writeLine('Finished "update-chars"', false);
 
@@ -160,7 +136,7 @@ class UpdateCharacters extends Command
                 }
 
                 if (! isset($names[$char->getId()]) || ! isset($affiliations[$char->getId()])) {
-                    $this->writeLine('  Character ' . $char->getId().': ' . self::UPDATE_NOK);
+                    $this->writeLine('  Character ' . $char->getId().': update NOK');
                     continue;
                 }
 
@@ -185,64 +161,10 @@ class UpdateCharacters extends Command
                 }
             }
             if (count($updateOk) > 0 && $this->entityManager->flush()) {
-                $this->writeLine('  Characters ' . implode(',', $updateOk).': ' . self::UPDATE_OK);
+                $this->writeLine('  Characters ' . implode(',', $updateOk).': update OK');
             }
 
             $this->entityManager->clear(); // detaches all objects from Doctrine
         } while (count($charIds) === $loopLimit);
-    }
-
-    private function updateCorps(): void
-    {
-        $offset = $this->dbResultLimit * -1;
-        do {
-            $offset += $this->dbResultLimit;
-            $corpIds = array_map(function (Corporation $corp) {
-                return $corp->getId();
-            }, $this->corpRepo->findBy([], ['lastUpdate' => 'ASC'], $this->dbResultLimit, $offset));
-
-            foreach ($corpIds as $corpId) {
-                if (! $this->entityManager->isOpen()) {
-                    $this->logger->critical('UpdateCharacters: cannot continue without an open entity manager.');
-                    break;
-                }
-                $this->entityManager->clear();
-                $this->checkErrorLimit();
-
-                $updatedCorp = $this->esiData->fetchCorporation($corpId);
-                if ($updatedCorp === null) {
-                    $this->writeLine('  Corporation ' . $corpId.': ' . self::UPDATE_NOK);
-                } else {
-                    $this->writeLine('  Corporation ' . $corpId.': ' . self::UPDATE_OK);
-                }
-
-                usleep($this->sleep * 1000);
-            }
-        } while (count($corpIds) === $this->dbResultLimit);
-    }
-
-    private function updateAlliances(): void
-    {
-        $alliIds = array_map(function (Alliance $alli) {
-            return $alli->getId();
-        }, $this->alliRepo->findBy([], ['lastUpdate' => 'ASC']));
-
-        foreach ($alliIds as $alliId) {
-            if (! $this->entityManager->isOpen()) {
-                $this->logger->critical('UpdateCharacters: cannot continue without an open entity manager.');
-                break;
-            }
-            $this->entityManager->clear();
-            $this->checkErrorLimit();
-
-            $updatedAlli = $this->esiData->fetchAlliance($alliId);
-            if ($updatedAlli === null) {
-                $this->writeLine('  Alliance ' . $alliId.': ' . self::UPDATE_NOK);
-            } else {
-                $this->writeLine('  Alliance ' . $alliId.': ' . self::UPDATE_OK);
-            }
-
-            usleep($this->sleep * 1000);
-        }
     }
 }